webapi/webapi-webrtc-w3c-tests/webrtc/RTCDataChannelEvent_channel_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu,Yun <yunx.liu@intel.com>

-->

<meta charset="utf-8">
<title>WebRTC Test: RTCDataChannelEvent - channel basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcdatachannelevent">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>

var t = async_test("Check if the readonly attribute channel of RTCDataChannelEvent exists and type of RTCDataChannel");

var atStep = 'Create offer';

var pc1 = new RTCPeerConnection(null);
var pc2 = new RTCPeerConnection(null);
var sendChannel = pc1.createDataChannel("This is pc1");

pc2.ondatachannel = t.step_func(function (event) {
  assert_true("channel" in event, "channel attribute in RTCDataChannelEvent");
  assert_class_string(event.channel, "RTCDataChannel", "channel attribute of type");
  assert_readonly(event, "channel", "expect attribute channel cannot be changed but");
  t.done();
});

pc1.onicecandidate = t.step_func(function(evt) {
  if(evt.candidate) {
    pc2.addIceCandidate(event.candidate);
  }
});

pc2.onicecandidate = t.step_func(function(evt) {
  if(evt.candidate) {
    pc1.addIceCandidate(event.candidate);
  }
});

pc1.createOffer()
.then(function(offer) {
  atStep = 'Set local description at one';
  return pc1.setLocalDescription(offer);
})
.then(function() {
  atStep = 'Set remote description at two';
  return pc2.setRemoteDescription(
    pc1.localDescription);
})
.then(function() {
  atStep = 'Create answer';
  return pc2.createAnswer();
})
.then(function(answer) {
  atStep = 'Set local description at two';
  return pc2.setLocalDescription(answer);
})
.then(function() {
  atStep = 'Set remote description at one';
  return pc1.setRemoteDescription(
    pc2.localDescription);
})
.then(function() {
  atStep = 'Negotiation completed';
})
.catch(t.step_func(function(e) {
  assert_unreached('Error ' + e.name + ': ' + e.message +
                   ' happened at step ' + atStep);
}));

</script>
